requirements.txt
# Deep Learning (PyTorch only)
torch>=2.0.0
torchvision>=0.15.0

# Medical Imaging
nibabel>=5.0.0
SimpleITK>=2.2.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Image Processing
opencv-python>=4.8.0
albumentations>=1.3.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# ML Utils
scikit-learn>=1.3.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Config & Jupyter
PyYAML>=6.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0